# src/main/resources/luckperms-openapi.yml
openapi: 3.0.0
info:
  title: LuckPerms REST API
  description: The LuckPerms REST API allows developers to query LuckPerms data using HTTP calls.
  version: '1.0'
  contact:
    name: LuckPerms Team
    url: 'https://luckperms.net'
  license:
    name: MIT
    url: 'https://github.com/LuckPerms/LuckPerms/blob/master/LICENSE.txt'
servers:
  - url: ../
...
paths:
  /user:
    get:
      summary: Get all existing users
      operationId: get-users
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/UniqueId'
              examples:
                example-1:
                  value:
                    - c1d60c50-70b5-4722-8057-87767557e50d
      description: Returns an array of all known users.
      tags:
        - Users
    post:
      summary: Create a new user
      operationId: create-user
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                example-1:
                  value:
                    uniqueId: c1d60c50-70b5-4722-8057-87767557e50d
                    username: Luck
                    nodes:
                      - key: group.default
                        type: inheritance
                        value: true
                        context: []
                    metadata:
                      meta: {}
                      primaryGroup: default
        '400':
          description: Missing required information
        '409':
          description: User already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
            examples:
              example-1:
                value:
                  uniqueId: c1d60c50-70b5-4722-8057-87767557e50d
                  username: Luck
        description: ''
      description: Create a new user.
      tags:
        - Users
  /user/search:
    get:
      summary: Search for users with given nodes
      tags:
        - Users
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSearchResult'
              examples:
                example-1:
                  value:
                    - uniqueId: c1d60c50-70b5-4722-8057-87767557e50d
                      results:
                        - key: minecraft.command.ban
                          type: permission
                          value: true
                          context:
                            - key: server
                              value: survival
                          expiry: 1658428395
      operationId: get-user-search
      description: |-
        Search for users with given nodes.

        You must specify one of the query parameters in the request.
      parameters:
        - schema:
            type: string
            minLength: 1
          in: query
          name: key
          description: Search for nodes with a key equal to
        - schema:
            type: string
            minLength: 1
          in: query
          name: keyStartsWith
          description: Search for nodes with a key starting with
        - schema:
            type: string
            minLength: 1
          in: query
          name: metaKey
          description: Search for meta nodes with a meta key equal to
        - schema:
            type: string
            enum:
              - regex_permission
              - inheritance
              - prefix
              - suffix
              - meta
              - weight
              - display_name
          in: query
          name: type
          description: Search for nodes with a type equal to
  '/user/{uniqueId}':
    parameters:
      - $ref: '#/components/parameters/uniqueId'
    get:
      summary: Get a users data
      tags:
        - Users
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                example-1:
                  value:
                    uniqueId: c1d60c50-70b5-4722-8057-87767557e50d
                    username: Luck
                    nodes:
                      - key: group.default
                        type: inheritance
                        value: true
                        context: []
                      - key: minecraft.command.ban
                        type: permission
                        value: true
                        context:
                          - key: server
                            value: survival
                        expiry: 1658428395
                    metadata:
                      meta:
                        nametag-color: red
                        fly-speed: '5'
                      prefix: '[Admin] '
                      suffix: ' [the great]'
                      primaryGroup: default
        '404':
          description: User doesn't exist
      operationId: get-user
      parameters:
        - $ref: '#/components/parameters/uniqueId'
      description: Get a user by unique id (UUID).
    patch:
      summary: Update a users data
      operationId: patch-user
      responses:
        '200':
          description: OK
        '400':
          description: Missing required information
        '404':
          description: User doesn't exist
      description: Update a user's data
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
                - properties:
                    username:
                      type: string
                      minLength: 1
                      example: Luck
              type: object
            examples:
              example-1:
                value:
                  username: Luck
        description: ''
      tags:
        - Users
    delete:
      summary: Delete a user
      operationId: delete-user
      responses:
        '200':
          description: OK
        '404':
          description: User doesn't exist
      description: Delete a user
      tags:
        - Users
...
  /group:
    get:
      summary: Get all existing groups
      operationId: get-groups
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/Name'
              examples:
                example-1:
                  value:
                    - admin
      description: Get all known groups
      tags:
        - Groups
    post:
      summary: Create a new group
      operationId: create-group
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              examples:
                example-1:
                  value:
                    name: admin
                    nodes: []
                    metadata:
                      meta: {}
        '400':
          description: Missing required information
        '409':
          description: Group already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroup'
            examples:
              example-1:
                value:
                  name: admin
        description: ''
      description: Create a new group.
      tags:
        - Groups
  /group/search:
    get:
      summary: Search for groups with given nodes
      tags:
        - Groups
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupSearchResult'
              examples:
                example-1:
                  value:
                    - name: admin
                      results:
                        - key: minecraft.command.ban
                          type: permission
                          value: true
                          context:
                            - key: server
                              value: survival
                          expiry: 1658428395
      operationId: get-group-search
      description: |-
        Search for groups with given nodes.

        You must specify one of the query parameters in the request.
      parameters:
        - schema:
            type: string
            minLength: 1
          in: query
          name: key
          description: Search for nodes with a key equal to
        - schema:
            type: string
            minLength: 1
          in: query
          name: keyStartsWith
          description: Search for nodes with a key starting with
        - schema:
            type: string
            minLength: 1
          in: query
          name: metaKey
          description: Search for meta nodes with a meta key equal to
        - schema:
            type: string
            enum:
              - regex_permission
              - inheritance
              - prefix
              - suffix
              - meta
              - weight
              - display_name
          in: query
          name: type
          description: Search for nodes with a type equal to
  '/group/{name}':
    parameters:
      - $ref: '#/components/parameters/name'
    get:
      summary: Get a groups data
      tags:
        - Groups
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              examples:
                example-1:
                  value:
                    name: admin
                    displayName: Admin
                    weight: 10
                    nodes:
                      - key: minecraft.command.ban
                        type: permission
                        value: true
                        context:
                          - key: server
                            value: survival
                        expiry: 1658428395
                    metadata:
                      meta:
                        nametag-color: red
                        fly-speed: '5'
                      prefix: '[Admin] '
                      suffix: ' [the great]'
                      primaryGroup: admin
        '404':
          description: Group doesn't exist
      operationId: get-group
      parameters:
        - $ref: '#/components/parameters/name'
      description: Get a group
    delete:
      summary: Delete a group
      operationId: delete-group
      responses:
        '200':
          description: OK
        '404':
          description: Group doesn't exist
      description: Delete a group
      tags:
        - Groups
...
  /action:
    post:
      summary: Submit a new action
      operationId: submit-action
      responses:
        '202':
          description: Accepted
        '400':
          description: Missing required information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Action'
      description: Submit a new action to the action logger.
      tags:
        - Actions
components:
  schemas:
    Node:
      type: object
      description: |-
        Represents a LuckPerms "node".

        Nodes encapsulate more than just permission assignments. Nodes are used to store data about inherited groups, as well as assigned prefixes, suffixes and meta values.
      title: ''
      properties:
        key:
          type: string
          minLength: 1
          example: minecraft.command.ban
        type:
          type: string
          enum:
            - permission
            - regex_permission
            - inheritance
            - prefix
            - suffix
            - meta
            - weight
            - display_name
        value:
          type: boolean
          default: true
        context:
          $ref: '#/components/schemas/ContextSet'
        expiry:
          type: integer
          example: 1658428395
      required:
        - key
        - type
        - value
        - context
...
    Context:
      type: object
      x-examples:
        example-1:
          key: server
          value: survival
      properties:
        key:
          type: string
          minLength: 1
          example: server
          pattern: '^[a-z0-9]+$'
        value:
          type: string
          minLength: 1
          example: survival
          pattern: '^[a-z0-9]+$'
      required:
        - key
        - value
      description: A context pair.
    ContextSet:
      title: ContextSet
      x-stoplight:
        id: x6gu6czcagtng
      type: array
      items:
        $ref: '#/components/schemas/Context'
      description: A set of context pairs.
      x-examples: {}
    Group:
      title: Group
      x-stoplight:
        id: 4xch4cn4mxc3u
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        displayName:
          type: string
          minLength: 1
          example: Admin
        weight:
          type: integer
        nodes:
          $ref: '#/components/schemas/NodeMap'
        metadata:
          $ref: '#/components/schemas/Metadata'
      required:
        - name
        - nodes
        - metadata
    NewGroup:
      title: NewGroup
      x-stoplight:
        id: 4xch4cn4mxc3u
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
      required:
        - name
    User:
      title: User
      x-stoplight:
        id: 4xch4cn4mxc3u
      type: object
      properties:
        uniqueId:
          $ref: '#/components/schemas/UniqueId'
        username:
          type: string
          minLength: 1
          example: Luck
        parentGroups:
          type: array
          items:
            $ref: '#/components/schemas/Name'
        nodes:
          $ref: '#/components/schemas/NodeMap'
        metadata:
          $ref: '#/components/schemas/Metadata'
      required:
        - uniqueId
        - nodes
        - metadata
    NewUser:
      title: NewUser
      x-stoplight:
        id: 4xch4cn4mxc3u
      type: object
      properties:
        uniqueId:
          $ref: '#/components/schemas/UniqueId'
        username:
          type: string
          minLength: 1
          example: Luck
      required:
        - uniqueId
        - username
    NodeMap:
      title: NodeMap
      x-stoplight:
        id: dlrdoescpnbhh
      type: array
      items:
        $ref: '#/components/schemas/Node'
      description: A map (array) of nodes.
    Metadata:
      title: Metadata
      x-stoplight:
        id: jweo4hdxmecgn
      type: object
      properties:
        meta:
          type: object
        prefix:
          type: string
          example: '[Admin] '
        suffix:
          type: string
          example: ' [the great]'
        primaryGroup:
          type: string
          example: default
          minLength: 1
      required:
        - meta
      x-examples:
        example-1:
          meta:
            nametag-color: red
            fly-speed: '5'
          prefix: '[Admin] '
          suffix: ' [the great]'
          primaryGroup: default
      description: Metadata about a user or group.
    UniqueId:
      type: string
      title: Unique Id
      x-stoplight:
        id: 1juihei0xo8qa
      description: A player unique id (UUID)
      example: c1d60c50-70b5-4722-8057-87767557e50d
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
    Name:
      type: string
      title: Name
      description: A group name
      example: admin
      pattern: '^[a-z0-9]+$'
      minLength: 1
...
    Action:
      title: Action
      x-stoplight:
        id: bsufd88gmrjvi
      type: object
      x-examples:
        example-1:
          timestamp: 1658428395
          source:
            uniqueId: c1d60c50-70b5-4722-8057-87767557e50d
            name: Luck
          target:
            uniqueId: 069a79f4-44e9-4726-a5be-fca90e38aaf5
            name: Notch
            type: user
          description: permission set minecraft.command.ban true
      description: ''
      properties:
        timestamp:
          type: integer
          description: 'The time when the action happened, defaults to the current time'
          example: 1658428395
        source:
          type: object
          description: Who performed the action
          required:
            - uniqueId
            - name
          properties:
            uniqueId:
              $ref: '#/components/schemas/UniqueId'
            name:
              type: string
              minLength: 1
              example: Luck
        target:
          type: object
          description: Who/what was the action performed on
          required:
            - name
            - type
          properties:
            uniqueId:
              $ref: '#/components/schemas/UniqueId'
            name:
              type: string
              example: Luck
            type:
              type: string
              enum:
                - user
                - group
                - track
        description:
          type: string
          minLength: 1
          example: permission set minecraft.command.ban true
          description: A description of the action
      required:
        - source
        - target
        - description
...
    UserSearchResult:
      title: UserSearchResult
      type: object
      properties:
        uniqueId:
          $ref: '#/components/schemas/UniqueId'
        results:
          $ref: '#/components/schemas/NodeMap'
      description: ''
    GroupSearchResult:
      title: GroupSearchResult
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        results:
          $ref: '#/components/schemas/NodeMap'
      description: ''
  securitySchemes:
    apikey:
      type: http
      scheme: bearer
      description: API Key used to authenticate
...
  parameters:
    uniqueId:
      name: uniqueId
      in: path
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
        example: c1d60c50-70b5-4722-8057-87767557e50d
      required: true
      description: A player unique id (UUID)
    name:
      name: name
      in: path
      schema:
        type: string
        example: admin
        pattern: '^[a-z0-9]+$'
        minLength: 1
      required: true
      description: A group name
security:
  - apikey: []